requirements.txt
asgiref==3.3.4
cloudinary==1.33.0
cryptography==3.4.8
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.5
Django==3.2.3
django-allauth==0.44.0
django-cart==1.0.4
django-crispy-forms==1.11.2
django-summernote==0.8.11.6
gunicorn==20.1.0
oauthlib==3.1.1
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.31.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
stripe==2.42.0
urllib3==1.26.18
